+++ src/vs/platform/userDataProfile/electron-sandbox/userDataProfile.ts
@@ -3,28 +3,40 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { Emitter } from 'vs/base/common/event';
+import { Disposable } from 'vs/base/common/lifecycle';
+import { joinPath } from 'vs/base/common/resources';
 import { UriDto } from 'vs/base/common/types';
+import { URI } from 'vs/base/common/uri';
 import { IChannel } from 'vs/base/parts/ipc/common/ipc';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
-import { IFileService } from 'vs/platform/files/common/files';
 import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/services';
-import { ILogService } from 'vs/platform/log/common/log';
-import { DidChangeProfilesEvent, IUserDataProfile, IUserDataProfilesService, reviveProfile, UseDefaultProfileFlags, UserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
+import { DidChangeProfilesEvent, IUserDataProfile, IUserDataProfilesService, reviveProfile, UseDefaultProfileFlags, WorkspaceIdentifier } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { ISingleFolderWorkspaceIdentifier, IWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
 
-export class UserDataProfilesNativeService extends UserDataProfilesService implements IUserDataProfilesService {
+export class UserDataProfilesNativeService extends Disposable implements IUserDataProfilesService {
+
+	readonly _serviceBrand: undefined;
 
 	private readonly channel: IChannel;
+
+	readonly profilesHome: URI;
+
+	get defaultProfile(): IUserDataProfile { return this.profiles[0]; }
+	private _profiles: IUserDataProfile[] = [];
+	get profiles(): IUserDataProfile[] { return this._profiles; }
+
+	private readonly _onDidChangeProfiles = this._register(new Emitter<DidChangeProfilesEvent>());
+	readonly onDidChangeProfiles = this._onDidChangeProfiles.event;
 
 	constructor(
 		profiles: UriDto<IUserDataProfile>[],
 		@IMainProcessService mainProcessService: IMainProcessService,
 		@IEnvironmentService environmentService: IEnvironmentService,
-		@IFileService fileService: IFileService,
-		@ILogService logService: ILogService,
 	) {
-		super(environmentService, fileService, logService);
+		super();
 		this.channel = mainProcessService.getChannel('userDataProfiles');
+		this.profilesHome = joinPath(environmentService.userRoamingDataHome, 'profiles');
 		this._profiles = profiles.map(profile => reviveProfile(profile, this.profilesHome.scheme));
 		this._register(this.channel.listen<DidChangeProfilesEvent>('onDidChangeProfiles')(e => {
 			const added = e.added.map(profile => reviveProfile(profile, this.profilesHome.scheme));
@@ -34,17 +46,19 @@ export class UserDataProfilesNativeService extends UserDataProfilesService imple
 		}));
 	}
 
-	override async createProfile(name: string, useDefaultFlags?: UseDefaultProfileFlags, workspaceIdentifier?: ISingleFolderWorkspaceIdentifier | IWorkspaceIdentifier): Promise<IUserDataProfile> {
+	async createProfile(name: string, useDefaultFlags?: UseDefaultProfileFlags, workspaceIdentifier?: ISingleFolderWorkspaceIdentifier | IWorkspaceIdentifier): Promise<IUserDataProfile> {
 		const result = await this.channel.call<UriDto<IUserDataProfile>>('createProfile', [name, useDefaultFlags, workspaceIdentifier]);
 		return reviveProfile(result, this.profilesHome.scheme);
 	}
 
-	override async setProfileForWorkspace(profile: IUserDataProfile, workspaceIdentifier: ISingleFolderWorkspaceIdentifier | IWorkspaceIdentifier): Promise<void> {
+	async setProfileForWorkspace(profile: IUserDataProfile, workspaceIdentifier: ISingleFolderWorkspaceIdentifier | IWorkspaceIdentifier): Promise<void> {
 		await this.channel.call<UriDto<IUserDataProfile>>('setProfileForWorkspace', [profile, workspaceIdentifier]);
 	}
 
-	override removeProfile(profile: IUserDataProfile): Promise<void> {
+	removeProfile(profile: IUserDataProfile): Promise<void> {
 		return this.channel.call('removeProfile', [profile]);
 	}
+
+	getProfile(workspaceIdentifier: WorkspaceIdentifier): IUserDataProfile { throw new Error('Not implemented'); }
 }
