+++ src/vs/platform/userDataProfile/test/common/userDataProfileService.test.ts
@@ -14,6 +14,7 @@ import { InMemoryFileSystemProvider } from 'vs/platform/files/common/inMemoryFil
 import { AbstractNativeEnvironmentService } from 'vs/platform/environment/common/environmentService';
 import product from 'vs/platform/product/common/product';
 import { UserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
+import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
 
 const ROOT = URI.file('tests').with({ scheme: 'vscode-tests' });
 
@@ -37,7 +38,7 @@ suite('UserDataProfileService (Common)', () => {
 		disposables.add(fileService.registerProvider(ROOT.scheme, fileSystemProvider));
 
 		environmentService = new TestEnvironmentService(joinPath(ROOT, 'User'));
-		testObject = new UserDataProfilesService(environmentService, fileService, logService);
+		testObject = new UserDataProfilesService(environmentService, fileService, new UriIdentityService(fileService), logService);
 	});
 
 	teardown(() => disposables.clear());
