+++ src/vs/workbench/services/extensions/test/browser/extensionService.test.ts
@@ -36,6 +36,8 @@ import { IExtensionHostManager } from 'vs/workbench/services/extensions/common/e
 import { IUserDataProfileService } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
 import { UserDataProfileService } from 'vs/workbench/services/userDataProfile/common/userDataProfileService';
 import { IUserDataProfilesService, UserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
+import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
+import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
 
 suite('BrowserExtensionService', () => {
 	test('pickRunningLocation', () => {
@@ -181,6 +183,7 @@ suite('ExtensionService', () => {
 			[IWorkspaceTrustEnablementService, WorkspaceTrustEnablementService],
 			[IUserDataProfilesService, UserDataProfilesService],
 			[IUserDataProfileService, UserDataProfileService],
+			[IUriIdentityService, UriIdentityService],
 		]);
 		extService = <MyTestExtensionService>instantiationService.get(IExtensionService);
 	});
